kafka-demo-3/data_backend/requirements.txt
kafka-python
fastapi
uvicorn
websockets
aiokafka